move plugin implementations to separate file, keep only registry in init (#83)
* move plugin implementations to separate file, keep only registry in init

* remove bibtex todo in wikitext benchmark
