+++ brainscore_language/data/fedorenko2016/__init__.py
@@ -16,6 +16,7 @@ BIBTEX = """@article{fedorenko2016neural,
   year={2016},
   publisher={National Acad Sciences}
 }"""
+
 data_registry['Fedorenko2016.language'] = lambda: load_from_s3(
     identifier="Fedorenko2016.language",
     version_id="qvB7YZfEjbXEE64bODNLlQlZKWGpgPhy",
